64bit: Fixed variables in CCDynamicArray.

+++ Common/csrun.cpp
@@ -182,7 +182,7 @@ struct CCDynamicArray : ICCDynamicObject
 
     // If it's an array of managed objects, release
     // their ref counts
-    long *elementCount = (long*)address;
+    int *elementCount = (int*)address;
     if (elementCount[0] & ARRAY_MANAGED_TYPE_FLAG)
     {
       elementCount[0] &= ~ARRAY_MANAGED_TYPE_FLAG;
@@ -202,8 +202,8 @@ struct CCDynamicArray : ICCDynamicObject
   // serialize the object into BUFFER (which is BUFSIZE bytes)
   // return number of bytes used
   virtual int Serialize(const char *address, char *buffer, int bufsize) {
-    long *sizeInBytes = &((long*)address)[-1];
-    long sizeToWrite = *sizeInBytes + 8;
+    int *sizeInBytes = &((int*)address)[-1];
+    int sizeToWrite = *sizeInBytes + 8;
     if (sizeToWrite > bufsize)
     {
       // buffer not big enough, ask for a bigger one
@@ -223,7 +223,7 @@ struct CCDynamicArray : ICCDynamicObject
   {
     char *newArray = new char[numElements * elementSize + 8];
     memset(newArray, 0, numElements * elementSize + 8);
-    long *sizePtr = (long*)newArray;
+    int *sizePtr = (int*)newArray;
     sizePtr[0] = numElements;
     sizePtr[1] = numElements * elementSize;
     if (isManagedType)
